Refactored file & folder management services to copy/move files in parrarel tasks

+++ MemoriesBackend.Core/Services/FileDatabaseService.cs
@@ -9,51 +9,68 @@ namespace MemoriesBackend.Application.Services
 {
     public class FileDatabaseService : IFileDatabaseService
     {
-        private readonly IGenericRepository<File> _fileRepository;
+        private readonly IFileRepository _fileRepository;
 
-        public FileDatabaseService(IGenericRepository<File> fileRepository)
+        public FileDatabaseService(IFileRepository fileRepository)
         {
             _fileRepository = fileRepository;
         }
 
         public async Task<File> CreateFileAsync(File file)
         {
+            if(file == null)
+            {
+                throw new ApplicationException("Failed to create - file is null");
+            }
+
             var createdFile = await _fileRepository.Create(file);
             return createdFile;
         }
 
-        public async Task DeleteFileAsync(Guid id)
+        public async Task DeleteFileAsync(Guid fileId)
         {
-            await _fileRepository.Delete(id);
-        }
+            var file = await _fileRepository.GetById(fileId);
+            if(file == null)
+            {
+                throw new ApplicationException($"Failed to delete - cannot find file with id {fileId}");
+            }
 
-        public async Task DeleteFileAsync(File file)
-        {
             _fileRepository.Delete(file);
         }
 
-        public async Task<File> GetFileByIdAsync(Guid id, bool asNoTracking = true)
+        public void DeleteFileAsync(File file)
         {
-            var file = await _fileRepository.GetById(id, asNoTracking);
+            if (file == null)
+            {
+                throw new ApplicationException($"Failed to delete - file is null");
+            }
+
+            _fileRepository.Delete(file);
+        }
+
+        public async Task<File> GetFileByIdAsync(Guid fileId, bool asNoTracking = true)
+        {
+            var file = await _fileRepository.GetById(fileId, asNoTracking);
             if (file == null) throw new ApplicationException("Failed to fetch - There was no file found with the given id.");
             return file;
         }
 
+        public async Task<File> GetFileByIdWithDetailsAsync(Guid fileId, bool asNoTracking = true)
+        {
+            return await _fileRepository.GetFileByIdWithDetailsAsync(fileId, asNoTracking);
+        }
+
         public async Task<IEnumerable<File>> GetAllFilesAsync(
+            Expression<Func<File, bool>> filter,
             int? pageNumber,
             int? pageSize,
-            Expression<Func<File, bool>>? filter = null,
             Func<IQueryable<File>, IOrderedQueryable<File>>? orderBy = null,
             bool asNoTracking = true
         )
         {
-            var files = await _fileRepository.GetAll(
-                filter: filter,
-                orderBy: orderBy,
-                pageNumber: pageNumber,
-                pageSize: pageSize,
-                asNoTracking: asNoTracking
-            ).ToListAsync();
+            var files = await _fileRepository
+                .GetAll(filter, orderBy, pageNumber, pageSize, asNoTracking)
+                .ToListAsync();
 
             return files;
         }
@@ -62,8 +79,18 @@ namespace MemoriesBackend.Application.Services
         {
             foreach (var file in files)
             {
-                await _fileRepository.Create(file);
+                await CreateFileAsync(file);
             }
+        }
+
+        public void UpdateFile(File file)
+        {
+            if(file == null)
+            {
+                throw new ApplicationException("Failed to update - file is null");
+            }
+
+            _fileRepository.Update(file);
         }
 
         public async Task SaveAsync()
